// gabinet-kosmetyczny-frontend/src/app/feature/clients/model/client.model.ts
export interface IClientData {
    accountUuid?: string;
    clientUuid?: string;
    name?: string;
    surname?: string;
    street?: string;
    code?: string;
    town?: string;
    birthday?: Date;
    phone?: string;
    email?: string;
    password?: string;
    role?: string;
    isActive?: boolean;
}